debug option を LazySync にする
