+++ lnb-common/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod config;
+pub mod debug;
 pub mod persistence;
 pub mod rate_limits;
 pub mod user_roles;
